Revert autostart icons, add checkbox for installing desktop files

+++ src/vorta/autostart.py
@@ -40,26 +40,19 @@ def open_app_at_startup(enabled=True):
                                "assets/metadata/com.borgbase.Vorta.desktop")) as desktop_file:
             desktop_file_text = desktop_file.read()
 
-            # Find XDG_CONFIG_HOME and XDG_DATA_HOME unless when running in flatpak
+            # Find XDG_CONFIG_HOME unless when running in flatpak
             if is_flatpak:
                 config_path = Path.home() / '.config'
-                data_path = Path.home() / ".local" / "share"
             else:
                 config_path = Path(os.environ.get(
                     "XDG_CONFIG_HOME", os.path.expanduser("~"))) / '.config'
-                data_path = Path(os.environ.get(
-                    "XDG_DATA_HOME", os.path.expanduser("~"))) / ".local" / "share"
 
             autostart_path = config_path / "autostart"
 
             autostart_file_path = autostart_path / 'vorta.desktop'
-            icon_path = data_path / "icons" / "hicolor" / "scalable" / "apps"
 
             if not autostart_path.exists():
                 autostart_path.mkdir(parents=True)
-
-            if not icon_path.exists():
-                icon_path.mkdir(parents=True)
 
             if enabled:
                 # Replace to for flatpak if appropriate and start in background
@@ -69,21 +62,47 @@ def open_app_at_startup(enabled=True):
                 # Add autostart delay
                 desktop_file_text += (AUTOSTART_DELAY)
 
-                # Copy icons
-                main_icon_path = Path(__file__).parent / "assets" / "metadata" / "com.borgbase.Vorta.svg"
-                symbolic_icon_path = Path(__file__).parent / "assets" / "metadata" / "com.borgbase.Vorta-symbolic.svg"
-                copy2(main_icon_path, icon_path)
-                copy2(symbolic_icon_path, icon_path)
-
                 # Write desktop file
                 autostart_file_path.write_text(desktop_file_text)
             else:
-                main_icon_path = icon_path / "com.borgbase.Vorta.svg"
-                symbolic_icon_path = icon_path / "com.borgbase.Vorta-symbolic.svg"
-
                 if autostart_file_path.exists():
                     autostart_file_path.unlink()
-                if symbolic_icon_path.exists():
-                    symbolic_icon_path.unlink()
-                if main_icon_path.exists():
-                    main_icon_path.unlink()
+
+
+def desktop_application(enabled=True):
+    is_flatpak = Path('/.flatpak-info').exists()
+
+    # Find XDG_DATA_HOME unless when running in flatpak
+    if is_flatpak:
+        data_path = Path.home() / ".local" / "share"
+    else:
+        data_path = Path(os.environ.get(
+            "XDG_DATA_HOME", os.path.expanduser("~"))) / ".local" / "share"
+
+    desktop_path = data_path / "applications"
+    icon_path = data_path / "icons" / "hicolor" / "scalable" / "apps"
+
+    if not icon_path.exists():
+        icon_path.mkdir(parents=True)
+    if not desktop_path.exists():
+        desktop_path.mkdir(parents=True)
+
+    if enabled:
+        # Copy icons
+        main_icon_path = Path(__file__).parent / "assets" / "metadata" / "com.borgbase.Vorta.svg"
+        symbolic_icon_path = Path(__file__).parent / "assets" / "metadata" / "com.borgbase.Vorta-symbolic.svg"
+        desktop_file_path = Path(__file__).parent / "assets" / "metadata" / "com.borgbase.Vorta.desktop"
+        copy2(main_icon_path, icon_path)
+        copy2(symbolic_icon_path, icon_path)
+        copy2(desktop_file_path, desktop_path)
+    else:
+        main_icon_path = icon_path / "com.borgbase.Vorta.svg"
+        symbolic_icon_path = icon_path / "com.borgbase.Vorta-symbolic.svg"
+        desktop_file_path = desktop_path / "com.borgbase.Vorta.desktop"
+
+        if desktop_file_path.exists():
+            desktop_file_path.unlink()
+        if symbolic_icon_path.exists():
+            symbolic_icon_path.unlink()
+        if main_icon_path.exists():
+            main_icon_path.unlink()
